VULCAN-18 Remove JAX-RS interfaces: Resource and classes who inherit from it
As a result of this we need to also remove ResourceMapper and ApiContributors and all the code depending on them. This functionality will be added in later commits as a single `Resource` class

It also provokes that `URIResolver` is no longer needed.

An finally, as a result of that all the `Optional.empty()` code is added temporary so the project still compile.